+++ src/lang/deprecated.js
@@ -4,6 +4,7 @@ import { Draggable } from "./../renderable/draggable.js";
 import { DropTarget } from "./../renderable/dragndrop.js";
 import UISpriteElement from "./../renderable/ui/uispriteelement.js";
 import { warning } from "./console.js";
+import CanvasRenderTarget from "../video/rendertarget/canvasrendertarget.js";
 
 /*
  * placeholder for all deprecated classes and corresponding alias for backward compatibility
@@ -141,4 +142,24 @@ Renderer.prototype.getHeight = function()  {
     return this.height;
 };
 
+/**
+ * @classdesc
+ * @deprecated since 17.1.0
+ * @see CanvasRenderTarget
+ */
+export class CanvasTexture extends CanvasRenderTarget {
+    /**
+     * @param {number} width - the desired width of the canvas
+     * @param {number} height - the desired height of the canvas
+     * @param {object} attributes - The attributes to create both the canvas and context
+     * @param {boolean} [attributes.context="2d"] - the context type to be created ("2d", "webgl", "webgl2")
+     * @param {boolean} [attributes.offscreenCanvas=false] - will create an offscreenCanvas if true instead of a standard canvas
+     * @param {boolean} [attributes.willReadFrequently=false] - Indicates whether or not a lot of read-back operations are planned
+     * @param {boolean} [attributes.antiAlias=false] - Whether to enable anti-aliasing, use false (default) for a pixelated effect.
+     */
+    constructor(width, height, attributes) {
+        warning("CanvasTexture", "CanvasRenderTarget", "17.1.0");
+        super(width, height, attributes);
+    }
+}