+++ src/particles/emitter.js
@@ -8,7 +8,7 @@ import Container from "./../renderable/container.js";
  * @ignore
  */
 function createDefaultParticleTexture(w = 8, h = 8) {
-    let defaultParticleTexture = pool.pull("CanvasTexture", w, h, { offscreenCanvas: true });
+    let defaultParticleTexture = pool.pull("CanvasRenderTarget", w, h, { offscreenCanvas: true });
 
     defaultParticleTexture.context.fillStyle = "#fff";
     defaultParticleTexture.context.fillRect(0, 0, w, h);