+++ src/renderable/light2d.js
@@ -101,7 +101,7 @@ export default class Light2d extends Renderable {
         this.visibleArea = pool.pull("Ellipse", this.centerX, this.centerY, this.width, this.height);
 
         /** @ignore */
-        this.texture = pool.pull("CanvasTexture", this.width, this.height, { offscreenCanvas: false });
+        this.texture = pool.pull("CanvasRenderTarget", this.width, this.height, { offscreenCanvas: false });
 
         this.anchorPoint.set(0, 0);